SIMPLY-3607: Removed LCP and Adept from OpenEbooks
SIMPLY-3607: Made Axis360 preferred acquisition in OpenEbooks

SIMPLY-3692: Added NYPLAxisAuthorizer
